Move authentication folder to the routes directory
