+++ server/package.json
@@ -15,7 +15,6 @@
   "dependencies": {
     "@nestjs/common": "^7.0.0",
     "@nestjs/core": "^7.0.0",
-    "@nestjs/jwt": "^7.2.0",
     "@nestjs/mongoose": "^7.1.2",
     "@nestjs/platform-express": "^7.0.0",
     "bcrypt": "^5.0.0",
