+++ server/src/app.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose/dist/mongoose.module';
-import { AuthModule } from './authentication/auth.module';
+import { AuthModule } from './routes/auth/auth.module';
 import config from './config';
-import { UsersModule } from './models/user/users.module';
+import { UsersModule } from './routes/user/users.module';
 
 @Module({
   imports: [
