+++ server/src/routes/auth/auth.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common";
-import { UsersModule } from "../models/user/users.module";
+import { UsersModule } from "../user/users.module";
 import { AuthController } from "./auth.controller";
 import { AuthService } from "./auth.service";
