+++ server/src/routes/auth/dto/register.dto.ts
