+++ server/src/routes/auth/schemas/register.schema.ts
