+++ server/src/routes/user/interfaces/IUser.ts
